Read game target, fix rarely happens looting stuck, delay multipler now can have decimal parts

+++ Multi Account Synchronizer/BotForm.cs
@@ -161,7 +161,7 @@ namespace Multi_Account_Synchronizer
 
         private void numericUpDown1_ValueChanged(object sender, EventArgs e)
         {
-            bot.DelayMultipler = (int)numericUpDown1.Value;
+            bot.DelayMultipler = (double)numericUpDown1.Value;
         }
     }
 }
